Move XML-building into separate service
There were a lot of private methods just being used to build this XML, so it felt like time to move it off to its own class.
